tests: add helpers file
